Change SubscriptionType from int to byte
Change type of SubscriptionType to match database type.

+++ backend/ResQueue/ResQueue/Dtos/Messages/SubscriptionDto.cs
@@ -4,6 +4,6 @@ public record SubscriptionDto(
     string TopicName,
     string DestinationType,
     string DestinationName,
-    int SubscriptionType,
+    byte SubscriptionType,
     string RoutingKey
 );
